fix: user_id field added.

+++ src/controllers/todo_controller.ts
@@ -1,7 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import { BadRequest, Conflict, NotFound } from 'http-errors';
-import { ValidationUtils } from '../utils';
+import { JWTUtils, ValidationUtils } from '../utils';
 import { Todo } from '../models';
+import { JwtPayload } from 'jsonwebtoken';
 
 const addTodo = async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -12,9 +13,10 @@ const addTodo = async (req: Request, res: Response, next: NextFunction) => {
     const doesExist = await Todo.findOne({ id: value.id });
     if (doesExist) throw Conflict('Id:' + value.id + ' is already exist');
 
-    const todo = new Todo(value);
-    const savedTodo = await todo.save();
-    console.log('Todo:', savedTodo);
+    const userId = reqToToken(req);
+
+    const todo = new Todo({ ...value, userId: userId });
+    await todo.save();
 
     res.status(200).json({
       message: 'success',
@@ -26,9 +28,13 @@ const addTodo = async (req: Request, res: Response, next: NextFunction) => {
 };
 
 const getTodos = async (req: Request, res: Response) => {
-  const activity = req.query.activity;
+  const userId = reqToToken(req);
 
-  const result = await Todo.find(activity ? { activity } : null);
+  // const queryActivity = req.query.activity;
+  // const activityInt = parseInt(queryActivity.toString());
+  // const activity = activityInt === 0 ? null : activityInt;
+
+  const result = await Todo.find({ userId: userId });
   const todos = result.map(value => {
     const todo = {
       id: value.id,
@@ -49,7 +55,9 @@ const deleteTodo = async (req: Request, res: Response, next: NextFunction) => {
     const id = req.params.id;
     if (!id) BadRequest();
 
-    const deletedTodo = await Todo.deleteOne({ id });
+    const userId = reqToToken(req);
+
+    const deletedTodo = await Todo.deleteOne({ id, userId });
     if (deletedTodo.deletedCount > 0) {
       res.status(200).json({ message: 'success', status: 200 });
     } else {
@@ -69,10 +77,18 @@ const updateTodo = async (req: Request, res: Response, next: NextFunction) => {
     const doesExist = await Todo.findOne({ id: value.id });
     if (!doesExist) throw Conflict('Id:' + value.id + ' is not already exist');
 
-    await Todo.findOneAndUpdate({ id: value.id }, value);
+    const userId = reqToToken(req);
+
+    await Todo.findOneAndUpdate({ id: value.id, userId: userId }, value);
     res.status(200).json({ message: 'success', status: 200 });
   } catch (error) {
     next(error);
   }
+};
+
+const reqToToken = (req: Request) => {
+  const token = req.headers.authorization.split(' ')[1];
+  const jwtPayload: JwtPayload = JWTUtils.verifyAccessToken(token);
+  return jwtPayload.aud;
 };
 export default { getTodos, addTodo, deleteTodo, updateTodo };
